src/github_import_data/requirements.txt
gql
python-dotenv
requests
requests_toolbelt
azure-cosmos
azure-identity